Reverting Base URL Changes and integrating Circle CI
Working Circle CI update + fixing a few additional broken links. The Base URL appears to be necessary to run locally as of this time as it crashes some pages and not others. Will look into ways of removing it when I integrate into Gitpages. For now just reverting it so that we can have a build that can run locally.